WIP - Improve date parsing code
We check if a time is midnight and treat it as if it doesn't exist if so. This approach ran into timezone issues.

There's another timezone issue where passing dates between the client and the server results in an off by one error where the first date of the results the client receives is yesterday instead of today.

This commit starts using a more robust library to hopefully eliminate timezone issues as a source of bugs moving forward.

We also add Jest in order to test some of the datetime parsing code.