WIP - Improve date parsing code
This commit continues the migration to luxon in an attempt to eradicate timezone issues.

The events enumeration logic is refactored to use the new library, as well as in preparation to test the code. However, jest doesn't play nicely with importing the yaml package.

In the process this also introduces a bug where my biweekly events get duplicated and end up with on the wrong day...