Change events logic to be based on a class (#2)
This removes the need for complex init and reset logic to handle global variables. That pattern of handling state was awkward at best. This change should make it easier to add support for programmatically adding events to the data file.

This also adds tests for the api route that is backed by the events logic.