Update watch tests following successful experiment
Since `chokidar` works on the server, and locally, we'll continue to use it even though it doesn't play well with the mocked file system.

This new test mocks chokidar itself, allowing us to verify that it's getting called, and probably working, despite not getting the code coverage from having actually tested it.